Move LZO to standalone target
