--- Sources/PartoutOpenVPN_ObjC/include/lzo/lzoconf.h
@@ -1,1 +1,0 @@
-../../lib/lzoconf.h
