--- Sources/PartoutOpenVPN_ObjC/include/lzo/lzodefs.h
@@ -1,1 +1,0 @@
-../../lib/lzodefs.h
