+++ vendors/lzo/COPYING
