+++ vendors/lzo/include/lzo/lzoconf.h
