+++ vendors/lzo/include/lzo/lzodefs.h
