+++ vendors/lzo/minilzo.c
